relabel comments and moves stuff out of tfn files
