+++ modular_darkpack/modules/vampire_the_masquerade/code/vampire_clan/trusted_whitelist.dm
