+++ tgstation.dme
@@ -478,6 +478,7 @@
 #include "code\__DEFINES\~darkpack\subsystems.dm"
 #include "code\__DEFINES\~darkpack\vampire_clan.dm"
 #include "code\__DEFINES\~darkpack\walls.dm"
+#include "code\__DEFINES\~darkpack\whitelist.dm"
 #include "code\__DEFINES\~darkpack\xp_and_freebie.dm"
 #include "code\__DEFINES\~darkpack\dcs\beastmaster_signals.dm"
 #include "code\__DEFINES\~darkpack\dcs\bouncer_signals.dm"
@@ -7989,6 +7990,7 @@
 #include "modular_darkpack\modules\vampire_the_masquerade\code\vampire_clan\clan_mark.dm"
 #include "modular_darkpack\modules\vampire_the_masquerade\code\vampire_clan\clan_mark_pref.dm"
 #include "modular_darkpack\modules\vampire_the_masquerade\code\vampire_clan\clan_pref.dm"
+#include "modular_darkpack\modules\vampire_the_masquerade\code\vampire_clan\trusted_whitelist.dm"
 #include "modular_darkpack\modules\vampire_the_masquerade\code\vampire_clan\clans\banu_haqim.dm"
 #include "modular_darkpack\modules\vampire_the_masquerade\code\vampire_clan\clans\brujah.dm"
 #include "modular_darkpack\modules\vampire_the_masquerade\code\vampire_clan\clans\caitiff.dm"
@@ -8127,10 +8129,11 @@
 #include "modular_darkpack\modules\werewolf_the_apocalypse\code\subsplats\tribes\garou.dm"
 #include "modular_darkpack\modules\werewolf_the_apocalypse\code\weapons\klaives.dm"
 #include "modular_darkpack\modules\westfield_mall\code\mall_areas.dm"
+#include "modular_darkpack\modules\whitelist\code\whitelist_enforcement.dm"
+#include "modular_darkpack\modules\whitelist\code\whitelist_savefile.dm"
+#include "modular_darkpack\modules\whitelist\code\whitelist_prefs.dm"
 #include "modular_darkpack\modules\westfield_mall\code\modular_rooms.dm"
 #include "modular_darkpack\modules\z_travel\code\helpers.dm"
 #include "modular_darkpack\modules\z_travel\code\manhole.dm"
 #include "modular_darkpack\modules\z_travel\code\transfer_point.dm"
-#include "modular_tfn\modules\powers\code\admin_discipline_editor.dm"
-#include "modular_tfn\modules\vampire_the_masquerade\code\vampire_clan\trusted_whitelist.dm"
 // END_INCLUDE
